+++ vue-3-todo/package.json
@@ -10,8 +10,8 @@
     "lint": "eslint . --fix"
   },
   "dependencies": {
+    "@fontsource/roboto": "5.2.5",
     "@mdi/font": "7.4.47",
-    "roboto-fontface": "*",
     "vue": "^3.5.13",
     "vuetify": "^3.8.1"
   },
