+++ vue-3-todo/src/layouts/view.vue
@@ -1,0 +1,29 @@
+<template>
+  <v-app id="inspire">
+    <v-navigation-drawer v-model="drawer">
+      <!--  -->
+    </v-navigation-drawer>
+
+    <v-app-bar>
+      <v-app-bar-nav-icon @click="drawer = !drawer"></v-app-bar-nav-icon>
+
+      <v-app-bar-title>Application</v-app-bar-title>
+    </v-app-bar>
+
+    <v-main>
+      <router-view></router-view>
+    </v-main>
+  </v-app>
+</template>
+
+<script setup>
+  import { ref } from 'vue'
+
+  const drawer = ref(null)
+</script>
+
+<script>
+  export default {
+    data: () => ({ drawer: null }),
+  }
+</script>
