+++ vue-3-todo/src/main.js
@@ -13,6 +13,9 @@ import App from './App.vue'
 // Composables
 import { createApp } from 'vue'
 
+// Styles
+import 'unfonts.css'
+
 const app = createApp(App)
 
 registerPlugins(app)
