+++ vue-3-todo/src/pages/Home.vue
@@ -1,3 +1,12 @@
 <template>
-    <h1>Olá Mundo!!!</h1>
+    <h1>Olá Mundo!!</h1>
 </template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style></style>
+
