+++ vue-3-todo/src/router/index.js
@@ -10,8 +10,8 @@ import { setupLayouts } from 'virtual:generated-layouts'
 import { routes } from 'vue-router/auto-routes'
 
 const router = createRouter({
-  history: createWebHistory(process.env.BASE_URL),
-  extendRoutes: (routes) => setupLayouts(router),
+  history: createWebHistory(import.meta.env.BASE_URL),
+  extendRoutes: (router) => setupLayouts(routes),
 })
 
 // Workaround for https://github.com/vitejs/vite/issues/11804
